pyproject.toml
[project]
name = "llm-driven-clinical-study-analysis"
version = "0.1.0"
description = "LLM-driven clinical study analysis"
authors = [
    {name = "Crazz-Zaac"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pyeuropepmc (>=1.14.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "pdfminer-six (>=20260107,<20260108)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "scispacy (>=0.6.2,<0.7.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
